Add script to publish Javadoc to GH Pages
The Javadocs are build via Gradle and pushed to the "gh-pages" branch
of this repository. The script is executed by Travis CI.

The variable $TRAVIS_BUILD_DIR is set by Travis CI, while $GIT_REPO_URL
must be provided as secure environment variable (via Travis CI).

No documentation is generated for pull requests, and to avoid that the
docs are generated four times, for each build target once, it is
further restricted to the "oraclejdk8" builds.