Run build and tests on native OS X via Travis CI
The library files for multiple OS are fetched as archive, whereby the
actual libraries were extracted from the Bitcoin Core 0.11 release.

Travis CI supports multiple plattforms (Ubuntu and OS X), which is
leveraged by a build matrix.

Unfortunally the JDK support on OS X still appears to be waggly, so
there is only one build on OS X, with the default JDK.

This change has no impact on the Javadoc deployment, which is triggered
for non-pull requests once.