Add ignored tests for script verification with BitcoinJ
Optionally the data driven script verification tests can also be used
to test the verification via BitcoinJ, however, this is currently
disabled, because BitcoinJ understands only a subset of the scripts
properly. In particular the support for certain (not consensus
critical) verification flags is mostly missing.

To test BitcoinJ, the test specification may be unignored manually.